refactor: app.js, homepage.js  폴더 위치 변경
